+++ cybox/test/objects/image_file_test.py
@@ -3,8 +3,10 @@
 
 import unittest
 
-from cybox.common import Hash, String
 from cybox.objects.image_file_object import ImageFile
+
+from cybox.common import Hash, String
+from cybox.compat import long
 import cybox.test
 from cybox.test import EntityTestCase
 from cybox.test.objects import ObjectTestCase
@@ -20,7 +22,7 @@ class TestImageFile(ObjectTestCase, unittest.TestCase):
         'image_width': 3000,
         'bits_per_pixel': 9000,
         'compression_algorithm': u"An algorithm",
-        
+
         'is_packed': False,
         'is_masqueraded': True,
         'file_name': u"example.txt",
@@ -29,7 +31,7 @@ class TestImageFile(ObjectTestCase, unittest.TestCase):
         'device_path': u"\\Device\\CdRom0",
         'full_path': u"C:\\Temp\\example.txt",
         'file_extension': u"txt",
-        'size_in_bytes': 1024L,
+        'size_in_bytes': long(1024),
         'magic_number': u"D0CF11E0",
         'file_format': u"ASCII Text",
         'hashes': [
